Rewrite to idiomatic TypeScript
I've tried to preserve the existing functionality where possible,
except for the `outputRenamingMap` option. Instead of writing a
renaming map to disk in a variety of formats, this now just calls a
callback with a JS object. This gives the caller more flexibility, and
the main benefit of writing to disk is compatibility with the Closure
compiler which isn't an issue for this package.